+++ src/ZODB/Connection.py
@@ -27,6 +27,7 @@ from persistent import PickleCache
 from persistent.interfaces import IPersistentDataManager
 from ZODB.interfaces import IConnection
 from ZODB.interfaces import IBlobStorage
+from ZODB.interfaces import IStorageTransactionMetaData
 from ZODB.blob import Blob, rename_or_copy_blob, remove_committed_dir
 from transaction.interfaces import ISavepointDataManager
 from transaction.interfaces import IDataManagerSavepoint
@@ -51,6 +52,7 @@ import six
 from .mvccadapter import HistoricalStorageAdapter
 
 from . import valuedoc
+from . import _compat
 
 global_reset_counter = 0
 
@@ -458,14 +460,21 @@ class Connection(ExportImport, object):
     def tpc_begin(self, transaction):
         """Begin commit of a transaction, starting the two-phase commit."""
         self._modified = []
+        meta_data = TransactionMetaData(
+            transaction.user,
+            transaction.description,
+            transaction.extension)
+        transaction.set_data(self, meta_data)
 
         # _creating is a list of oids of new objects, which is used to
         # remove them from the cache if a transaction aborts.
         self._creating.clear()
-        self._normal_storage.tpc_begin(transaction)
+        self._normal_storage.tpc_begin(meta_data)
 
     def commit(self, transaction):
         """Commit changes to an object"""
+
+        transaction = transaction.data(self)
 
         if self._savepoint_storage is not None:
 
@@ -611,6 +620,8 @@ class Connection(ExportImport, object):
                 obj._p_serial = s
 
     def tpc_abort(self, transaction):
+        transaction = transaction.data(self)
+
         if self._import:
             self._import = None
 
@@ -667,6 +678,9 @@ class Connection(ExportImport, object):
             vote = self._storage.tpc_vote
         except AttributeError:
             return
+
+        transaction = transaction.data(self)
+
         try:
             s = vote(transaction)
         except ReadConflictError as v:
@@ -683,6 +697,8 @@ class Connection(ExportImport, object):
     def tpc_finish(self, transaction):
         """Indicate confirmation that the transaction is done.
         """
+        transaction = transaction.data(self)
+
         serial = self._storage.tpc_finish(transaction)
         assert type(serial) is bytes, repr(serial)
         for oid_iterator in self._modified, self._creating:
@@ -1270,3 +1286,38 @@ large_record_size option of the ZODB.DB constructor (or the
 large-record-size option in a configuration file) to specify a larger
 size.
 """
+
+@implementer(IStorageTransactionMetaData)
+class TransactionMetaData(object):
+
+    def __init__(self, user=u'', description=u'', extension=b''):
+        if not isinstance(user, bytes):
+            user = user.encode('utf-8')
+        self.user = user
+
+        if not isinstance(description, bytes):
+            description = description.encode('utf-8')
+        self.description = description
+
+        if not isinstance(extension, dict):
+            extension = _compat.loads(extension) if extension else {}
+        self.extension = extension
+
+    def note(self, text): # for tests
+        text = text.strip()
+        if not isinstance(text, bytes):
+            text = text.encode('utf-8')
+        if self.description:
+            self.description = self.description.strip() + b' ' + text
+        else:
+            self.description = text
+
+    @property
+    def _extension(self):
+        warnings.warn("_extension is deprecated, use extension",
+                      DeprecationWarning)
+        return self.extension
+
+    @_extension.setter
+    def _extension(self, v):
+        self.extension = v