+++ src/ZODB/DB.py
@@ -24,7 +24,7 @@ from . import utils
 
 from ZODB.broken import find_global
 from ZODB.utils import z64
-from ZODB.Connection import Connection
+from ZODB.Connection import Connection, TransactionMetaData
 from ZODB._compat import Pickler, _protocol, BytesIO
 import ZODB.serialize
 
@@ -469,7 +469,7 @@ class DB(object):
         self.large_record_size = large_record_size
 
         # Make sure we have a root:
-        with self.transaction('initial database creation') as conn:
+        with self.transaction(u'initial database creation') as conn:
             try:
                 conn.get(z64)
             except KeyError:
@@ -901,7 +901,7 @@ class DB(object):
 
         See :meth:`ZODB.interfaces.IStorage.history`.
         """
-        return self.storage.history(oid, size)
+        return _text_transaction_info(self.storage.history(oid, size))
 
     def supportsUndo(self):
         """Return whether the database supports undo.
@@ -920,7 +920,7 @@ class DB(object):
 
         if not self.supportsUndo():
             return ()
-        return self.storage.undoLog(*args, **kw)
+        return _text_transaction_info(self.storage.undoLog(*args, **kw))
 
     def undoInfo(self, *args, **kw):
         """Return a sequence of descriptions for transactions.
@@ -929,7 +929,7 @@ class DB(object):
         """
         if not self.supportsUndo():
             return ()
-        return self.storage.undoInfo(*args, **kw)
+        return _text_transaction_info(self.storage.undoInfo(*args, **kw))
 
     def undoMultiple(self, ids, txn=None):
         """Undo multiple transactions identified by ids.
@@ -1037,19 +1037,28 @@ class TransactionalUndo(object):
         pass
 
     def tpc_begin(self, transaction):
-        self._storage.tpc_begin(transaction)
+        tdata = TransactionMetaData(
+            transaction.user,
+            transaction.description,
+            transaction.extension)
+        transaction.set_data(self, tdata)
+        self._storage.tpc_begin(tdata)
 
     def commit(self, transaction):
+        transaction = transaction.data(self)
         for tid in self._tids:
             self._storage.undo(tid, transaction)
 
     def tpc_vote(self, transaction):
+        transaction = transaction.data(self)
         self._storage.tpc_vote(transaction)
 
     def tpc_finish(self, transaction):
+        transaction = transaction.data(self)
         self._storage.tpc_finish(transaction)
 
     def tpc_abort(self, transaction):
+        transaction = transaction.data(self)
         self._storage.tpc_abort(transaction)
 
     def sortKey(self):
@@ -1064,3 +1073,12 @@ def connection(*args, **kw):
     managing a separate database object.
     """
     return DB(*args, **kw).open_then_close_db_when_connection_closes()
+
+_transaction_meta_data_text_variables = 'user_name', 'description'
+def _text_transaction_info(info):
+    for d in info:
+        for name in _transaction_meta_data_text_variables:
+            if name in d:
+                d[name] = d[name].decode('utf-8')
+
+    return info